Fix records with annotations
